ActivityId må være long

+++ sandvika-activity-board-backend/src/main/java/no/bouvet/sandvika/activityboard/api/ActivityController.java
@@ -122,7 +122,7 @@ public class ActivityController {
     }
 
     @RequestMapping(value = "/activities/{activityId}/pointcalculation", method = RequestMethod.GET)
-    public PointsCalculation getPointsCalculationForActivity(@PathVariable("activityId") int activityId) {
+    public PointsCalculation getPointsCalculationForActivity(@PathVariable("activityId") long activityId) {
         return activityUtils.getPointsCalculationForActivity(activityId);
     }
 }
